Data from Finkel et al. (2012) evaluating Kenya's NCEP II-Uraia

This is a modified dataset from Finkel, Horowitz, and Rojo-Mendoza (2012) examining the effect of the NCEP II-Uraia, a civic engagemnet program ocurring in Kenya in 2007. The treatment ("treat") is participation in at least one of the specified civic engagement activities before the 2007 election. Several outcomes were examined in the original paper; here only political knowledge ("polknow") is included.

This dataset was used in King and Nielsen (2019) as an example demonstrating the propensity score mathcing paradox.

data("finkel2012")

Format

A data frame with 3141 observations on the following 16 variables.

treat

integer: treatment (0: untreated, 1: treated)

uraiamedia

integer; Uraia media exposure (0: No, 1: Yes)

age

integer; age (years)

churchgo

integer; frequency of church attendence (0: never, ..., 5: more than once a week)

groupactive

numeric; group activity scale, from 0 (no memberships) to 2 (extremely active)

income

integer; income level (0: up to kshs 5000, ..., 4: 15,000+)

groupleader

integer; leader of any group (0: No, 1: Yes)

male

integer; sex (0: female, 1: male)

educ

integer; education level (0: no formal schooling, ..., 9: university)

religion

factor: religion ("protestant", "catholic", "muslim", "other")

media

numeric: media exposure, from 1 (never) to 4 (daily)

polinterest

numeric: poltical interest, from 1 (very little interest) to 3 (a great deal of interest)

poldiscuss

integer: frequency of polticial discussion in the last year (0: never, ..., 2: several times)

civicgroup

integer: participation in a civic group (0: not a member, ..., 2: an active member)

polparty

integer: participation in a political party (0: not a member, ..., 2: an active member)

polknow

integer: political knowledge (0: minimum knowledge, ..., 4: maximum knowledge)

Details

Although many of the included variables are ordinal in nature, here they are analyzed as numeric values.

References

Finkel, S. E., Horowitz, J., & Rojo-Mendoza, R. T. (2012). Civic Education and Democratic Backsliding in the Wake of Kenya's Post-2007 Election Violence. The Journal of Politics. doi:10.1017/S0022381611001162

King, G., & Nielsen, R. (2019). Why Propensity Scores Should Not Be Used for Matching. Political Analysis, 1-20. doi:10.1017/pan.2019.11