Summarize a frontierEstimates object

summary() displays the effect estimates, confidence bounds, and, if present, model dependence bounds at user-specified points along the forntier.

# S3 method for frontierEstimates
summary(object, N, Ndrop, ...)

# S3 method for summary.frontierEstimates
print(x, digits, ...)

Arguments

object

a frontierEstimates object; the output of a call to estimateEffects().

N

the number of units remaining in the sample at the desired point(s) on the frontier. Exactly one of N or Ndrop can be specified. If both are omitted, will display effect estimates at the first, last, and middle point along the frontier.

Ndrop

the number of units dropped in the sample at the desired point(s) on the frontier. Exactly one of N or Ndrop can be specified. If both are omitted, will display effect estimates at the first, last, and middle point along the frontier.

x

a summary.frontierEstimates object; the output of a call to summary.frontierEstimates().

digits

the number of significant digits to display the estimates. Passed to format().

...

ignored.

Details

summary() simply extracts the effect estimates and confidence nad model dependence bounds from the frontierEstimates objec at the requested pointst; no further calculation is performed. Whenever the number of points along the frontier for which the effect is estimated in the call to estimateEffects() is less than the total number of points on the frontier, it is possible that the specified points on the frontier do not have effects estimated, in which case the estimates at the nearest points will be displayed.

Value

A summary.frontierEstimates object, which contains the following components:

Est

the effect estimates at the user-specified points on the frontier.

CIlower

the confidence interval lower bound at the user-specified points on the frontier.

CIupper

the confidence interval upper bound at the user-specified points on the frontier.

CIlevel

the percentiles corresponding to the confidence level.

MDlower

the model dependence lower bound at the user-specified points on the frontier if requested in the call to estimateEffects().

MDlower

the model dependence upper bound at the user-specified points on the frontier if requested in the call to estimateEffects().

See also

Examples

#See example at help("estimateEffects")