Summarize a matchFrontier object

summary() displays the sample sizes and imbalance metrics at the beginning of the frontier, the end of the frontier, and the point at which the imbalance metric is lowest, or at user-specified points along the frontier.

# S3 method for matchFrontier
summary(object, N, Ndrop, ...)

# S3 method for summary.matchFrontier
print(x, digits, ...)

Arguments

object

a matchFrontier object; the output of a call to makeFrontier().

N

the number of units remaining in the sample at the desired point(s) on the frontier. Exactly one of N or Ndrop can be specified.

Ndrop

the number of units dropped in the sample at the desired point(s) on the frontier. Exactly one of N or Ndrop can be specified.

x

a summary.matchFrontier object; the output of a call to summary.matchFrontier().

digits

the number of significant digits to display the imbalance metric. Passed to format().

...

ignored.

Details

The lowest value of the imbalance metric may not occur at the end of the frontier (i.e., with the smallest remaining sample size). summary() can be used to complement plot.matchFrontier() to identify which sample size yields the least imbalance or to view the values of the imbalance metric at individual points along the frontier.

When object is a MatchItFrontier object and propensity score matching was done in the original call to matchit(), an additiona column, Caliper, will be displayed. This contains the largest distance between members of each of the remaining pairs at the given remaining sample size, corresponding to the value of the caliper (in standardized units) that would need to be set to arrive at a sample with the given size. (Note that this value is obtained from the frontier, and may not correspond to the caliper actually required to achieve the desired sample size when matching, unless matching with replacement.)

A carrot (i.e., ^) is printed below the column the corresponds to the n component of the matchFrontier object. It is this value that the N and Ndrop arguments refer to in summary.matchFrontier(), frontier_to_matchit(), and generateDataset(). For example, if the carrot is pointing at the N control column, then supplying a value of 50 to N in generateDataset() will produce a dataset with 50 control units (and however many treated units are present).

Note that when no point on the frontier corresponds to the supplied value of N or Ndrop, the nearest point will be used.

Value

A summary.matchFrontier object, which contains the following components:

Ntreated

the number of treated units at the start, end, and best point on the frontier (best meaning with the lowest imbalance), or at the user-specified points on the frontier.

Ncontrol

the number of control units at the start, end, and best point on the frontier, or at the user-specified points on the frontier.

N

the total number of units at the start, end, and best point on the frontier, or at the user-specified points on the frontier.

Stat

the value of the imbalance metric at the start, end, and best point on the frontier, or at the user-specified points on the frontier.

bestind

the index of the best point on the frontier. NULL when N or Ndrop are specified.

QOI

the quantity of interest; the argument to QOI supplied in the original call to makeFrontier().

metric

the imbalance metric; the argument to metric supplied in the original call to makeFrontier().

caliper

when object is a MatchItFrontier object and propensity score matching was done in the original call to matchit(), the value of the caliper corresponding to the given sample size at the start, end, and best point on the frontier, or at the user-specified points on the frontier.

See also

Examples

# See examples at help("makeFrontier") and
# help("makeFrontier.matchit")